gens update

+++ Tests/RNGLRAbstractParser.Tests/gen.sh
@@ -1,9 +1,6 @@
 mono ../../Bin/Release/v40/YC.AbstractLexer.Generator.exe Lexer.fsl -o Lexer.fs
 
-if [-f log.exe ]
-then
-    rm log.txt
-fi
+[ -f log.txt ] && rm log.txt
 
 for i in AandB SimpleCalc Calc EpsilonKiller simpleCalc_with_Nterms simpleCalc_with_Nterms_2 simpleCalc_with_Nterms_3 simpleCalc_with_Nterms_4 PrettySimpleCalc NotAmbigousSimpleCalc NotAmbigousSimpleCalcWith2Ops Stars Stars2 Eps List FirstEps CroppedBrackets Brackets Brackets1 StrangeBrackets ErrorSupport
 do
